Add files via upload
Added code to keep worm from steering backwards over itself.
Tried to add worm_death function to make something happen when the worm touches the edge of the screen. For now, it just turns the head white, but I am trying to draw a larger ellipse at the spot where the head hit the edge.